Feat: added the new 'default_traits = MyType: bool' and 'derive = MyType: Trait' options to the make_types! macro.
Feat: Code Documentation updated
