+++ lambda-appsync-proc/doc/appsync_operation.md
@@ -6,17 +6,15 @@ type matching what is defined in the schema. The function will be wired up to ha
 for that operation through the AWS AppSync Direct Lambda resolver.
 
 # Important
-This macro can only be used in a crate where the [appsync_lambda_main!] macro has been used at the
-root level (typically in `main.rs`). The code generated by this macro depends on types and
-implementations that are created by [appsync_lambda_main!].
+This macro can only be used in a crate where [make_operation!] (or [make_appsync!]) has been invoked
+at the root level (typically in `main.rs`). The code generated by this macro depends on types and
+implementations that are created by [make_operation!] (or [make_appsync!], which calls it internally).
+The legacy [appsync_lambda_main!] macro (behind the `compat` feature) also works.
 
 # Example Usage
 
-```no_run
-# lambda_appsync::appsync_lambda_main!(
-#    "schema.graphql",
-#     exclude_lambda_handler = true,
-# );
+```rust,no_run
+# lambda_appsync::make_appsync!("schema.graphql");
 # mod sub {
 # async fn dynamodb_get_players() -> Result<Vec<Player>, AppsyncError> {
 #    todo!()
@@ -26,7 +24,7 @@ implementations that are created by [appsync_lambda_main!].
 # }
 use lambda_appsync::{appsync_operation, AppsyncError};
 
-// Your types are declared at the crate level by the appsync_lambda_main! macro
+// Your types are declared at the crate level by the make_appsync! macro
 use crate::Player;
 
 // Execute when a 'players' query is received
@@ -50,18 +48,15 @@ async fn create_player(name: String) -> Result<Player, AppsyncError> {
 You may need to explore the [AppsyncEvent](struct.AppsyncEvent.html) received by the lambda
 in your operation handler. You can make it available by adding the `with_appsync_event` flag and
 adding a reference to it in your operation handler signature (must be the last argument), like so:
-```no_run
-# lambda_appsync::appsync_lambda_main!(
-#    "schema.graphql",
-#     exclude_lambda_handler = true,
-# );
+```rust,no_run
+# lambda_appsync::make_appsync!("schema.graphql");
 # mod sub {
 # async fn dynamodb_create_player(name: String) -> Result<Player, AppsyncError> {
 #    todo!()
 # }
 use lambda_appsync::{appsync_operation, AppsyncError, AppsyncEvent, AppsyncIdentity};
 
-// Your types are declared at the crate level by the appsync_lambda_main! macro
+// Your types are declared at the crate level by the make_appsync! macro
 use crate::{Operation, Player};
 
 // Use the AppsyncEvent
@@ -87,15 +82,12 @@ the argument values for the operation.
 
 By default the [macro@appsync_operation] macro will discard your function's name but
 you can also keep it available by adding the `keep_original_function_name` flag:
-```no_run
-# lambda_appsync::appsync_lambda_main!(
-#    "schema.graphql",
-#     exclude_lambda_handler = true,
-# );
+```rust,no_run
+# lambda_appsync::make_appsync!("schema.graphql");
 # mod sub {
 use lambda_appsync::{appsync_operation, AppsyncError};
 
-// Your types are declared at the crate level by the appsync_lambda_main! macro
+// Your types are declared at the crate level by the make_appsync! macro
 use crate::Player;
 
 # async fn dynamodb_get_players() -> Result<Vec<Player>, AppsyncError> {
@@ -116,11 +108,9 @@ async fn other_stuff() {
 
 ## Using enhanced subscription filters
 
-```no_run
-# lambda_appsync::appsync_lambda_main!(
-#    "schema.graphql",
-#     exclude_lambda_handler = true,
-# );
+```rust,no_run
+# lambda_appsync::make_appsync!("schema.graphql");
+
 // (Optional) Use an enhanced subscription filter for onCreatePlayer
 use lambda_appsync::{appsync_operation, AppsyncError};
 use lambda_appsync::subscription_filters::{FilterGroup, Filter, FieldPath};
@@ -138,13 +128,11 @@ async fn on_create_player(name: String) -> Result<Option<FilterGroup>, AppsyncEr
 
 When using a single [FieldPath](subscription_filters/struct.FieldPath.html) you can turn it directly into a [FilterGroup](subscription_filters/struct.FilterGroup.html).
 The following code is equivalent to the one above:
-```no_run
-# lambda_appsync::appsync_lambda_main!(
-#    "schema.graphql",
-#     exclude_lambda_handler = true,
-# );
+```rust,no_run
+# lambda_appsync::make_appsync!("schema.graphql");
 # use lambda_appsync::{appsync_operation, AppsyncError};
 # use lambda_appsync::subscription_filters::{FilterGroup, Filter, FieldPath};
+
 #[appsync_operation(subscription(onCreatePlayer))]
 async fn on_create_player(name: String) -> Result<Option<FilterGroup>, AppsyncError> {
     Ok(Some(FieldPath::new("name")?.contains(name).into()))
