+++ lambda-appsync-proc/src/internal/make_appsync/legacy/mod.rs
@@ -1,1 +1,3 @@
+//! Legacy macro implementations retained for backwards compatibility.
+
 pub mod appsync_lambda_main;
