+++ lambda-appsync-proc/src/internal/mod.rs
@@ -1,3 +1,5 @@
+//! Internal implementation modules for the proc-macro crate.
+
 pub(crate) mod appsync_operation;
 pub(crate) mod common;
 pub(crate) mod make_appsync;
